Fix bug where customer has a active order and user tries to delete this customer

+++ src/main/java/seedu/address/logic/commands/DeleteCustomerCommand.java
@@ -40,6 +40,9 @@ public class DeleteCustomerCommand extends DeleteCommand {
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+        if (personToDelete.getOrders().size() != 0) {
+            throw new CommandException(Messages.MESSAGE_CUSTOMER_HAS_ORDER);
+        }
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
     }
